# hyveman-web/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "hyveman-api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/hosts": {
      "get": {
        "tags": [
          "Hosts"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Hosts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/HostInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/HostInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HostDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hosts/{id}": {
      "get": {
        "tags": [
          "Hosts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HostDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostDetailDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Hosts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/HostInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/HostInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HostDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Hosts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "confirm",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/hosts/{id}/idrac-cert": {
      "delete": {
        "tags": [
          "Hosts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/hosts/{id}/vms": {
      "get": {
        "tags": [
          "Hosts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VmDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VmDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VmDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "hostId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sourceId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "severityMin",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "eventId",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int64"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventSearchResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSearchResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSearchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/saved-searches": {
      "get": {
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedSearchDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedSearchDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedSearchDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Events"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/saved-searches/{id}": {
      "patch": {
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/registration-tokens": {
      "post": {
        "tags": [
          "Sources"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationTokenCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationTokenCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationTokenCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationTokenCreatedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationTokenCreatedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationTokenCreatedDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Sources"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistrationTokenDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistrationTokenDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistrationTokenDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registration-tokens/{id}/revoke": {
      "post": {
        "tags": [
          "Sources"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/sources/{sourceId}/tokens/{tokenId}/revoke": {
      "post": {
        "tags": [
          "Sources"
        ],
        "parameters": [
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tokenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventDto": {
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "sourceId": {
            "type": "string"
          },
          "sourceName": {
            "type": [
              "null",
              "string"
            ]
          },
          "hostId": {
            "type": [
              "null",
              "string"
            ]
          },
          "hostName": {
            "type": [
              "null",
              "string"
            ]
          },
          "dedupScope": {
            "type": "string"
          },
          "recordId": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "severity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "facility": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "channel": {
            "type": [
              "null",
              "string"
            ]
          },
          "eventId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "task": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "opcode": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "keywords": {
            "type": [
              "null",
              "string"
            ]
          },
          "fieldsJson": {
            "type": [
              "null",
              "string"
            ]
          },
          "rawJson": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "EventSearchResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
      "HostDetailDto": {
        "type": "object",
        "properties": {
          "rollupState": {
            "type": "string"
          },
          "rollupAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "components": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComponentDto"
            }
          },
          "latestMetrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MetricDto"
            }
          },
          "recentAlerts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertDto"
            }
          },
          "recentEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            }
          },
          "agent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentStatusDto"
              }
            ]
          },
          "idracStatus": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdracStatusDto"
              }
            ]
          },
          "idracCert": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IdracCertPinDto"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "sourceId": {
            "type": [
              "null",
              "string"
            ]
          },
          "idracUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "idracCredentialSet": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "HostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "sourceId": {
            "type": [
              "null",
              "string"
            ]
          },
          "idracUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "idracCredentialSet": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "HostInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "kind": {
            "type": [
              "null",
              "string"
            ]
          },
          "sourceId": {
            "type": [
              "null",
              "string"
            ]
          },
          "idracUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "idracUsername": {
            "type": [
              "null",
              "string"
            ]
          },
          "idracPassword": {
            "type": [
              "null",
              "string"
            ]
          },
          "enabled": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "confirmDelete": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
      "SavedSearchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "filter": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SavedSearchInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "filter": {
            "type": [
              "null",
              "object"
            ]
          },
          "updatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "VmDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "heartbeatOk": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "cpuPct": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "memMb": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "replicationState": {
            "type": [
              "null",
              "string"
            ]
          },
          "replicationHealth": {
            "type": [
              "null",
              "string"
            ]
          },
          "replicationLastApplyTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastSeen": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "stale": {
            "type": "boolean"
          }
        }
      }
    }
  },
...
}